Corrected soullichen. Several bugfixes.

+++ src/main/java/com/chipoodle/devilrpg/client/render/entity/layer/WerewolfArrowLayer.java
@@ -29,8 +29,8 @@ public class WerewolfArrowLayer<T extends LivingEntity, M extends WerewolfTransf
     protected void renderStuckItem(PoseStack p_116569_, MultiBufferSource p_116570_, int p_116571_, Entity p_116572_, float p_116573_, float p_116574_, float p_116575_, float p_116576_) {
         float f = Mth.sqrt(p_116573_ * p_116573_ + p_116575_ * p_116575_);
         Arrow arrow = new Arrow(p_116572_.level, p_116572_.getX(), p_116572_.getY(), p_116572_.getZ());
-        arrow.setYRot((float) (Math.atan2((double) p_116573_, (double) p_116575_) * (double) (180F / (float) Math.PI)));
-        arrow.setXRot((float) (Math.atan2((double) p_116574_, (double) f) * (double) (180F / (float) Math.PI)));
+        arrow.setYRot((float) (Math.atan2(p_116573_, p_116575_) * (double) (180F / (float) Math.PI)));
+        arrow.setXRot((float) (Math.atan2(p_116574_, f) * (double) (180F / (float) Math.PI)));
         arrow.yRotO = arrow.getYRot();
         arrow.xRotO = arrow.getXRot();
         this.dispatcher.render(arrow, 0.0D, 0.0D, 0.0D, 0.0F, p_116576_, p_116569_, p_116570_, p_116571_);
